setup maneuvering node
