+++ njord_tasks/config/params.yaml
@@ -6,3 +6,7 @@ task:
     start_lon: 149.1653442
     finish_lat: -35.3632621
     finish_lon: 149.1652374
+maneuvering:
+  ros__parameters:
+    distance_to_move: 10.0 # meters
+    angle_from_buoys: 10.0 # degrees
