+++ njord_tasks/include/njord_tasks/maneuvering_component.hpp
@@ -3,13 +3,10 @@
 
 #include <rclcpp/rclcpp.hpp>
 #include <std_msgs/msg/int32.hpp>
-#include <std_msgs/msg/float64.hpp>
-#include <chrono>
-#include <mavros_msgs/msg/state.hpp>
-#include "sensor_msgs/msg/nav_sat_fix.hpp"
-#include "geographic_msgs/msg/geo_pose_stamped.hpp"
+#include "njord_tasks_interfaces/msg/start_task.hpp"
+
 using std::placeholders::_1;
-using namespace std::chrono_literals;
+
 namespace njord_tasks
 {
 
@@ -19,43 +16,27 @@ public:
     explicit Maneuvering(const rclcpp::NodeOptions & options);
 
 private:
+    void taskToExecuteCallback(const njord_tasks_interfaces::msg::StartTask::SharedPtr msg);
     rcl_interfaces::msg::SetParametersResult param_callback(const std::vector<rclcpp::Parameter> &params);
-    void stateCallback(const mavros_msgs::msg::State::SharedPtr msg);
-    void timerCallback();
-    void poseCallback(const geographic_msgs::msg::GeoPoseStamped::SharedPtr msg);
-    rclcpp::Subscription<mavros_msgs::msg::State>::SharedPtr state_sub_;
+
+    rclcpp::Subscription<njord_tasks_interfaces::msg::StartTask>::SharedPtr task_to_execute_sub_;
+    rclcpp::Publisher<std_msgs::msg::Int32>::SharedPtr task_completion_status_pub_;
     rclcpp::node_interfaces::OnSetParametersCallbackHandle::SharedPtr on_set_parameters_callback_handle_;
-    rclcpp::TimerBase::SharedPtr timer_;
-    rclcpp::Subscription<geographic_msgs::msg::GeoPoseStamped>::SharedPtr pose_sub_;
 
-    void maneuver();
-
-    
-    int    p_wait_time_;
-    double p_speed_factor_;
-    double p_global_wp_reached_rad_;
-    double p_start_lat_;
-    double p_start_lon_;
-    double p_finish_lat_;
-    double p_finish_lon_;
-    
-    geographic_msgs::msg::GeoPoseStamped current_global_pose_;
-    bool in_guided_;
-
-    enum States {WAIT_FOR_GUIDED, WAIT_TO_REACH_START, MANEUVER, WAIT_TO_REACH_FINISH, COMPLETE}; 
-    States status_;
+    double p_distance_to_move_;
+    double p_angle_from_buoys_;
 
     template <typename T>
     void getParam(std::string param_name, T& param, T default_value, std::string desc)
     {
-     auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
-     param_desc.description = desc;
-     this->declare_parameter<T>(param_name, default_value, param_desc);
-     this->get_parameter(param_name, param);
-     std::string param_log_output = param_name + ": " + std::to_string(param);
-     RCLCPP_INFO(this->get_logger(), param_log_output.c_str()); 
+      auto param_desc = rcl_interfaces::msg::ParameterDescriptor{};
+      param_desc.description = desc;
+      this->declare_parameter<T>(param_name, default_value, param_desc);
+      this->get_parameter(param_name, param);
+      std::string param_log_output = param_name + ": " + std::to_string(param);
+      RCLCPP_INFO(this->get_logger(), param_log_output.c_str()); 
 
-     return;
+      return;
     }
 };
